ARM64: dts: msm8953: Correctly configure DWC3 for kernel 4.14
On kernel 4.14 the USB DBM was merged inside of the dwc3-msm
platform driver, since it was being used *only* by MSM: remove
the node and move the qcom,reset-ep-after-lpm-resume property
to the DWC3 MSM node.

Also, add linux,sysdev_is_parent to the standard Synopsis DWC3
driver, as the real DWC device is parent of it, plus declare
that this is USB core ID 0, our maximum speed and OTG mode.